+++ CHANGELOG.rst
@@ -2,10 +2,15 @@ Changelog for Mantarray Frontend Components
 ===========================================
 
 1.4.5 (unreleased)
+------------------
 
 Added:
 ^^^^^^
 - Status alert in top right of screen for if stimulation is running.
+
+Changed:
+^^^^^^^^
+- Labels in Platemap Editor will be universal to all platemaps intead of unique to one platemap
 
 
 1.4.4 (2023-03-30)